storage: start using sessionStorage for the communication with cockpit-storage
In release 311 cockpit-storaged started using sessionStorage for
communication with anaconda. Introduce this and also temporarilly
keep the data in localStorage as well. This allows us to maintain backwards
compatibility with older cockpit-storage releases. We need this for the purpose
of not breaking installability tests, by requesting the not-everywhere available
cockpit-storaged version in the spec file.